feat: add new deploy.js and deploy-erc20.js for openzeppelin upgradeable contracts
