+++ eth-contracts/test/data/geth-genesis.json
@@ -17,7 +17,12 @@
     "0x876fDe0F545ACf26ad75D5065370EFb4DFDC46C2": { "balance": "0x99999999999999999999" },
     "0x2D7E77f647ed9321B5BdB9F3235f5f981BdCC2aF": { "balance": "0x99999999999999999999" },
     "0x46beaC96B726a51C5703f99eC787ce12793Dae11": { "balance": "0x99999999999999999999" },
-    "0xC2F2D954Bb6296b923BC938c32C4C30A8e39015f": { "balance": "0x999999999999999999" },
+    "0xC2F2D954Bb6296b923BC938c32C4C30A8e39015f": { "balance": "0x99999999999999999999" },
+    "0x40f2a96F337D041c7aC8f4F6c548d25e6eC6d619": { "balance": "0x99999999999999999999" },
+    "0xb3896bc01B0552d29b1ea4F3997fbD13865a9104": { "balance": "0x99999999999999999999" },
+    "0x9104bD0fcd2C86F81407F6aa098182cc7621b6d0": { "balance": "0x99999999999999999999" },
+    "0x8951a3DdEf2bB36fF3846C3B6968812C269f4561": { "balance": "0x99999999999999999999" },
+    "0x42e8763917A72e07369AD54B158b0FA839f060bc": { "balance": "0x99999999999999999999" },
     "0x0000000000000000000000000000000000000009": { "balance": "0x01" }
   },
   "difficulty":"2",
